refactor: Fold snapshots.py into runtime.py in Daytona provider
Merge snapshot management helpers (alist_snapshots, aget_snapshot,
acreate_snapshot, aresolve_snapshot) into runtime.py. Delete snapshots.py.
Update all import references in sandbox_common.py and tests.

Net: -1 file, Daytona provider 13 → 12 files.
